# frontend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignupDto" }
            }
          }
        },
        "responses": { "201": { "description": "OTP sent to email" } },
        "summary": "Register new user and send OTP",
        "tags": ["Authentication"]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyEmailDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponseDto" }
              }
            }
          }
        },
        "summary": "Verify email with OTP",
        "tags": ["Authentication"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponseDto" }
              }
            }
          }
        },
        "summary": "Login with email and password",
        "tags": ["Authentication"]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "cookie": [] }],
        "summary": "Refresh access token using cookie",
        "tags": ["Authentication"]
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ForgotPasswordDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "summary": "Send password reset token",
        "tags": ["Authentication"]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "summary": "Reset password using reset token",
        "tags": ["Authentication"]
      }
    },
    "/auth/google": {
      "post": {
        "operationId": "AuthController_googleLoginOrSignup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GoogleLoginDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponseDto" }
              }
            }
          }
        },
        "summary": "Login or signup using Google ID token",
        "tags": ["Authentication"]
      }
    },
...
  },
  "info": {
    "title": "Survix API",
    "description": "Backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "user@email.com" },
          "username": {
            "type": "string",
            "example": "john_doe",
            "minLength": 3,
            "maxLength": 30
          },
          "password": {
            "type": "string",
            "description": "User password",
            "minLength": 8,
            "maxLength": 64,
            "example": "StrongPassword123!"
          }
        },
        "required": ["email", "username", "password"]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "user@email.com" },
          "otp": {
            "type": "string",
            "example": "123456",
            "description": "6-digit OTP code"
          }
        },
        "required": ["email", "otp"]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "60f7f0b3b4d1c2a5f8e9a123" },
          "email": { "type": "string", "example": "user@email.com" },
          "username": { "type": "object", "example": "john_doe" },
          "name": { "type": "object", "example": "John Doe" },
          "avatar": {
            "type": "object",
            "example": "https://cdn.app.com/avatar.png"
          }
        },
        "required": ["id", "email"]
      },
      "AuthTokensDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token"
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh token (httpOnly cookie is preferred)"
          },
          "accessTokenExpiresIn": {
            "type": "number",
            "description": "Access token expiration in seconds",
            "example": 900
          },
          "refreshTokenExpiresIn": {
            "type": "number",
            "description": "Refresh token expiration in seconds",
            "example": 604800
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "accessTokenExpiresIn",
          "refreshTokenExpiresIn"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/UserResponseDto" },
          "tokens": { "$ref": "#/components/schemas/AuthTokensDto" }
        },
        "required": ["user", "tokens"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "user@email.com" },
          "password": { "type": "string", "example": "StrongPassword123!" }
        },
        "required": ["email", "password"]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "user@email.com" }
        },
        "required": ["email"]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID (Mongo ObjectId)",
            "example": "60f7f0b3b4d1c2a5f8e9a123"
          },
          "token": {
            "type": "string",
            "description": "Reset token sent via email"
          },
          "newPassword": {
            "type": "string",
            "description": "User password",
            "minLength": 8,
            "maxLength": 64,
            "example": "StrongPassword123!"
          }
        },
        "required": ["userId", "token", "newPassword"]
      },
      "GoogleLoginDto": {
        "type": "object",
        "properties": {
          "googleToken": {
            "type": "string",
            "description": "Google ID token obtained from client-side Google auth"
          }
        },
        "required": ["googleToken"]
      },
...
    }
  }
}